Remove redundant ModelStub class declarations, minor refactoring
